Implement upload of file chunks
The code now uploads all of the chunks to object storage. Sadly,
it panics on the last chunk because it writes more bytes than it
read out of the file. I'll work on that soon.

Additionally, it does not yet write the manifest file to object
storage, so there isn't an SLO yet. Just ton of chunks of an SLO.