Implement single read, hash & upload stage
The new ReadHashAndUpload stage has a tiny memory footprint, but is
as powerful as the previous ReadData, HashData, and UploadData stages
that used enourmous quantities of memory. It should replace using those
stages unless there is some feature of it that causes problems for you
use-case. If that is the case, write custom upload code templated from
ReadHashAndUpload.