Restructure package hierarchy
We decided that it made more sense for the DloUploader and the
SloUploader to live side-by-side in the basic swiftlygo package than it
did for you to have to import special functionality to get the slo
uploader. Now DLO and SLO uploads are both available in the base
package, and the low-level pipeline api for SLOs is available in the
pipeline subpackage.